requirements.txt
customtkinter
Pillow
whois
requests